requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dateutils==0.6.12
dj-database-url==2.3.0
Django==5.1.3
django-autocomplete-light==3.12.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
